chore: seperate cpponent for MiniGameInstructions
